+++ src/components/layout/Sidebar.tsx
@@ -5,7 +5,7 @@ import { LINKS } from "../utils/constants/link";
 const MENU = [
   { icon: <HomeIcon />, label: "ホーム", href: LINKS.HOME },
   { icon: <MemberIcon />, label: "メンバー", href: LINKS.MEMBER },
-  { icon: <TextIcon />, label: "教材", href: LINKS.TEXT },
+  { icon: <TextIcon />, label: "教材", href: LINKS.TEACHINGMATERIAL },
   { icon: <StudyMeetingIcon />, label: "勉強会", href: LINKS.STUDYMEETING },
   { icon: <UsersIcon />, label: "1on1", href: LINKS.ONEONONE },
 ];
