+++ src/components/page/Carender/index.tsx
@@ -1,14 +1,13 @@
-import { Suspense } from "react";
+import { FC, Suspense } from "react";
 import { Layout } from "src/components/layout";
 import { AppLoading } from "src/components/ui-libraries/AppLoading";
 
-const OneonOne = () => {
+export const Calender: FC = () => {
   return (
     <Layout>
       <Suspense fallback={<AppLoading />}>
-        <h1>1on1</h1>
+        <h1>Calendar</h1>
       </Suspense>
     </Layout>
   );
 };
-export default OneonOne;
