+++ src/components/page/OneOnOne/index.tsx
@@ -1,14 +1,13 @@
-import { Suspense } from "react";
+import { FC, Suspense } from "react";
 import { Layout } from "src/components/layout";
 import { AppLoading } from "src/components/ui-libraries/AppLoading";
 
-const Text = () => {
+export const OneOnOne: FC = () => {
   return (
     <Layout>
       <Suspense fallback={<AppLoading />}>
-        <h1>教材</h1>
+        <h1>1on1</h1>
       </Suspense>
     </Layout>
   );
 };
-export default Text;
