+++ src/components/page/SignUp/index.tsx
@@ -1,0 +1,10 @@
+import { FC } from "react";
+
+export const SignUp: FC = () => {
+  return (
+    <div className="flex justify-center items-center h-screen">
+      <p>初回ログイン時には、このページに飛ばします</p>
+      <p>ここで、登録が完了したら、rootにredirectするようにする</p>
+    </div>
+  );
+};
