--- src/components/pages/Home/index.tsx
@@ -1,10 +1,0 @@
-import { Button } from "@mantine/core";
-
-export const Home = () => {
-  return (
-    <div className="font-sans text-red-500 text-clip">
-      <Button>Hello</Button>
-      Hello World
-    </div>
-  );
-};
