--- src/components/pages/mypage/index.tsx
@@ -1,7 +1,0 @@
-export const mypage = () => {
-  return (
-    <>
-      <h1>mypage</h1>
-    </>
-  );
-};
