+++ src/pages/calendar/index.tsx
@@ -1,14 +1,8 @@
-import { Suspense } from "react";
-import { Layout } from "src/components/layout";
-import { AppLoading } from "src/components/ui-libraries/AppLoading";
+import { NextPage } from "next";
+import { Calender } from "src/components/page/Carender";
 
-const Calendar = () => {
-  return (
-    <Layout>
-      <Suspense fallback={<AppLoading />}>
-        <h1>Calendar</h1>
-      </Suspense>
-    </Layout>
-  );
+const Calendar: NextPage = () => {
+  return <Calender />;
 };
+
 export default Calendar;
