+++ src/pages/one-on-one/index.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from "next";
+import { OneOnOne } from "src/components/page/OneOnOne";
+
+const OneOnOnePage: NextPage = () => {
+  return <OneOnOne />;
+};
+
+export default OneOnOnePage;
