+++ src/pages/signup/index.tsx
@@ -1,14 +1,10 @@
+import { SignUp } from "src/components/page/SignUp";
 import type { NextPage } from "next";
 
-const SignUp: NextPage = () => {
-  return (
-    <div className="flex justify-center items-center h-screen">
-      <p>初回ログイン時には、このページに飛ばします</p>
-      <p>ここで、登録が完了したら、rootにredirectするようにする</p>
-    </div>
-  );
+const SignUpPage: NextPage = () => {
+  return <SignUp />;
 };
 
-export default SignUp;
+export default SignUpPage;
 
 // todo: LoginButtonを共通化してそっから持ってくる
