+++ src/pages/teaching-material/index.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from "next";
+import { TeachingMaterial } from "src/components/page/TeachingMaterial";
+
+const TeachingMaterialPage: NextPage = () => {
+  return <TeachingMaterial />;
+};
+
+export default TeachingMaterialPage;
